legend finally works but lots of  flashing going on
