+++ src/app.js
@@ -5,7 +5,6 @@ import { ObservationDialog } from "@components/dialog/observation-dialog";
 import { useLayers } from '@context';
 import { Sidebar } from '@components/sidebar';
 import { ControlPanel } from '@components/control-panel';
-import { MapLegend } from '@components/legend';
 
 /**
  * renders the main content
@@ -33,7 +32,6 @@ const Content = () => {
             <Map />
             <Sidebar />
             <ControlPanel/>
-            <MapLegend />
         </Fragment>
     );
 };
