+++ src/utils/default-styles.js
@@ -1,6 +1,3 @@
-//export const maxeleStyle = '<?xml version="1.0" encoding="UTF-8"?> \
-  //<ogc:GetMap xmlns:ogc="http://www.opengis.net/ows" xmlns:gml="http://www.opengis.net/gml" version="1.1.1" service="WMS"> \
-//'<?xml version="1.0" encoding="UTF-8"?><sld:StyledLayerDescriptor xmlns="http://www.opengis.net/sld" xmlns:sld="http://www.opengis.net/sld" xmlns:gml="http://www.opengis.net/gml" xmlns:ogc="http://www.opengis.net/ogc" version="1.0.0"> \
 export const maxeleStyle = '<?xml version="1.0" encoding="UTF-8"?> \
   <sld:StyledLayerDescriptor xmlns="http://www.opengis.net/sld" xmlns:sld="http://www.opengis.net/sld" xmlns:gml="http://www.opengis.net/gml" xmlns:ogc="http://www.opengis.net/ogc" version="1.1.0"> \
     <sld:NamedLayer> \
@@ -20,22 +17,21 @@ export const maxeleStyle = '<?xml version="1.0" encoding="UTF-8"?> \
               </sld:GrayChannel> \
             </sld:ChannelSelection> \
             <sld:ColorMap> \
-              <sld:ColorMapEntry color="#0000CC" quantity="0.00" label="0.00 m"/> \
-              <sld:ColorMapEntry color="#0000FF" quantity="0.1875" label="0.19 m"/> \
-              <sld:ColorMapEntry color="#0033FF" quantity="0.375" label="0.38 m"/> \
-              <sld:ColorMapEntry color="#0066FF" quantity="0.5625" label="0.56 m"/> \
-              <sld:ColorMapEntry color="#0099FF" quantity="0.75" label="0.75 m"/> \
-              <sld:ColorMapEntry color="#00CCFF" quantity="0.9375" label="0.94 m"/> \
-              <sld:ColorMapEntry color="#00FFFF" quantity="1.125" label="1.13 m"/> \
-              <sld:ColorMapEntry color="#33FFCC" quantity="1.3125" label="1.31 m"/> \
-              <sld:ColorMapEntry color="#66FF99" quantity="1.50" label="1.50 m"/> \
-              <sld:ColorMapEntry color="#99FF66" quantity="1.6875" label="1.69 m"/> \
-              <sld:ColorMapEntry color="#CCFF33" quantity="1.875" label="1.88 m"/> \
-              <sld:ColorMapEntry color="#FFFF00" quantity="2.0625" label="2.06 m"/> \
-              <sld:ColorMapEntry color="#FFCC00" quantity="2.25" label="2.25 m"/> \
-              <sld:ColorMapEntry color="#FF9900" quantity="2.4375" label="2.44 m"/> \
-              <sld:ColorMapEntry color="#FF6600" quantity="2.625" label="2.63 m"/> \
-              <sld:ColorMapEntry color="#FF3300" quantity="2.8125" label="2.81 m"/> \
+              <sld:ColorMapEntry color="#0000FF" quantity="0.00" label="0.00 m"/> \
+              <sld:ColorMapEntry color="#0033FF" quantity="0.20" label="0.20 m"/> \
+              <sld:ColorMapEntry color="#0066FF" quantity="0.40" label="0.40 m"/> \
+              <sld:ColorMapEntry color="#0099FF" quantity="0.60" label="0.60 m"/> \
+              <sld:ColorMapEntry color="#00CCFF" quantity="0.80" label="0.80 m"/> \
+              <sld:ColorMapEntry color="#00FFFF" quantity="1.00" label="1.00 m"/> \
+              <sld:ColorMapEntry color="#33FFCC" quantity="1.20" label="1.20 m"/> \
+              <sld:ColorMapEntry color="#66FF99" quantity="1.40" label="1.40 m"/> \
+              <sld:ColorMapEntry color="#99FF66" quantity="1.60" label="1.60 m"/> \
+              <sld:ColorMapEntry color="#CCFF33" quantity="1.80" label="1.80 m"/> \
+              <sld:ColorMapEntry color="#FFFF00" quantity="2.00" label="2.00 m"/> \
+              <sld:ColorMapEntry color="#FFCC00" quantity="2.20" label="2.20 m"/> \
+              <sld:ColorMapEntry color="#FF9900" quantity="2.40" label="2.40 m"/> \
+              <sld:ColorMapEntry color="#FF6600" quantity="2.60" label="2.60 m"/> \
+              <sld:ColorMapEntry color="#FF3300" quantity="2.80" label="2.80 m"/> \
               <sld:ColorMapEntry color="#FF0000" quantity="3.00" label="3.00 m"/> \
             </sld:ColorMap> \
             <sld:ContrastEnhancement/> \
